Move tests to ParseFormulaTest
